requirements.txt
accelerate 
beautifulsoup4
bertopic
chardet
config
ekphrasis
google-colab
hdbscan
html5lib
langdetect
MarkupSafe
matplotlib
mpmath
nltk
numpy
openai==0.28.1
pandas
plotly
python-dotenv
pytorch-lightning
scikit-learn
seaborn
sentence-transformers
tensorboard
torch
transformers
tweet-preprocessor
umap-learn
wordcloud
